Fix using screenshot step with response cache
Change the output of the `Screenshot` step, from an array `['response'
=> RespondedRequest, 'screenshotPath' => string]` to a
`RespondedRequestWithScreenshot` object, that has a `screenshotPath`
property. The problem with the previous solution was: when using the
response cache, the step failed, because it gets a cached response from
the loader that was not actually loaded in the headless browser. When
the step afterwards tries to take a screenshot from the page that is
still open in the browser, it just fails because there is no open page.
Now, with the new `RespondedRequestWithScreenshot` object, the
`screenshotPath` is also saved in the cached response.